refactor: move enums.ts to @/config/
