+++ src/app/guide_articles/layout.tsx
@@ -4,19 +4,21 @@ import { usePathname } from "next/navigation";
 import type { ReactNode } from "react";
 
 import { AbstractResourceLayout } from "@/components/abstract/abstract-resource-layout";
-import { Resource } from "@/lib/enums";
+import { Resource } from "@/config/enums";
 
 export default function Layout({ children }: { children: ReactNode }) {
   const pathname = usePathname();
 
-  const titleMap = {
-    [`/${Resource.GuideArticles}/create`]: "Dodawanie artykułu",
-    [`/${Resource.GuideArticles}/edit`]: "Edycja artykułu",
-    [`/${Resource.GuideArticles}`]: "Zarządzenie artykułami",
-  };
-
   return (
-    <AbstractResourceLayout titleMap={titleMap} pathname={pathname}>
+    <AbstractResourceLayout
+      resource={Resource.GuideArticles}
+      titleMap={{
+        "/create": "Dodawanie artykułu",
+        "/edit": "Edycja artykułu",
+        "": "Zarządzenie artykułami",
+      }}
+      pathname={pathname}
+    >
       {children}
     </AbstractResourceLayout>
   );
