+++ src/app/student_organizations/layout.tsx
@@ -4,19 +4,21 @@ import { usePathname } from "next/navigation";
 import type { ReactNode } from "react";
 
 import { AbstractResourceLayout } from "@/components/abstract/abstract-resource-layout";
-import { Resource } from "@/lib/enums";
+import { Resource } from "@/config/enums";
 
 export default function Layout({ children }: { children: ReactNode }) {
   const pathname = usePathname();
 
-  const titleMap = {
-    [`/${Resource.StudentOrganizations}/create`]: "Dodawanie organizacji",
-    [`/${Resource.StudentOrganizations}/edit`]: "Edycja organizacji",
-    [`/${Resource.StudentOrganizations}`]: "Zarządzanie organizacjami",
-  };
-
   return (
-    <AbstractResourceLayout titleMap={titleMap} pathname={pathname}>
+    <AbstractResourceLayout
+      resource={Resource.StudentOrganizations}
+      titleMap={{
+        "/create": "Dodawanie organizacji",
+        "/edit": "Edycja organizacji",
+        "": "Zarządzanie organizacjami",
+      }}
+      pathname={pathname}
+    >
       {children}
     </AbstractResourceLayout>
   );
