+++ src/components/abstract/abstract-resource-edit-page.tsx
@@ -1,7 +1,7 @@
 import type { ComponentType } from "react";
 
 import { RESOURCE_API_PATHS } from "@/config/constants";
-import type { Resource } from "@/lib/enums";
+import type { Resource } from "@/config/enums";
 import { fetchQuery } from "@/lib/fetch-utils";
 import { sanitizeId } from "@/lib/helpers";
 import type { ResourceTypes } from "@/types/app";
