+++ src/lib/error-handling.ts
@@ -1,10 +1,10 @@
-import { ERROR_MESSAGES } from "@/config/constants";
+import { API_ERROR_MESSAGES } from "@/config/constants";
 
 import { FetchError } from "./fetch-utils";
 
 function getCodedMessage(code: string, fallback: string): string {
-  if (code in ERROR_MESSAGES) {
-    return ERROR_MESSAGES[code];
+  if (code in API_ERROR_MESSAGES) {
+    return API_ERROR_MESSAGES[code];
   }
   console.warn(
     `Unhandled error code: ${code}. Please add it to src/error-handling.ts#ERROR_MESSAGES with a user-friendly error message.`,
@@ -14,7 +14,7 @@ function getCodedMessage(code: string, fallback: string): string {
 
 export function getErrorMessage(
   error: unknown,
-  fallback: string = ERROR_MESSAGES.E_UNEXPECTED_ERROR,
+  fallback: string = API_ERROR_MESSAGES.E_UNEXPECTED_ERROR,
 ): string {
   if (error instanceof FetchError) {
     const code = error.errorReport?.error.code;
